+++ spec/lib/concierge/suppliers/waytostay/properties.rb
@@ -218,10 +218,77 @@ RSpec.shared_examples "Waytostay property client" do
 
   end
 
+  describe "#add_late_checkin_fees" do
+    let(:property) { Roomorama::Property.new('test') }
+    let(:response) {
+      Concierge::SafeAccessHash.new(
+        "general" => {
+          "late_checkin_possible" => true,
+          "late_checkin_from" => "22:00",
+          "late_checkin_to" => "08:00",
+          "late_checkin_fee" => "30"
+        },
+        "payment" => {
+          "currency" => "USD"
+        }
+      )
+    }
+
+    before {
+      property.description_append ||= ""
+      property.de.description_append ||= ""
+      property.es.description_append ||= ""
+      property.zh.description_append ||= ""
+      property.zh_tw.description_append ||= ""
+
+      described_class.new.send(:add_late_checkin_fee, property, response)
+    }
+
+    it "should be correct" do
+      expect(property.description_append).to eq "Late check-in fees from 22:00 to 08:00 - 30 USD\n"
+      expect(property.de.description_append).to eq "Gebühr in Höhe von 30 USD für späte Anreise (22:00 bis 08:00)\n"
+      expect(property.es.description_append).to eq "Tarde de facturación de 22:00 a 08:00 - 30 USD\n"
+      expect(property.zh.description_append).to eq "延迟入住费从22:00至08:00  -  30 USD\n"
+      expect(property.zh_tw.description_append).to eq "延遲入住費從22:00至08:00  -  30 USD\n"
+    end
+
+    context "late checkin not possible" do
+      let(:response) {
+        Concierge::SafeAccessHash.new(
+          "general" => {
+            "late_checkin_possible" => false,
+            "late_checkin_from" => "22:00",
+            "late_checkin_to" => "08:00",
+            "late_checkin_fee" => "30"
+          },
+          "payment" => {
+            "currency" => "USD"
+          }
+        )
+      }
+
+      it "should not append to description_apend" do
+        expect(property.description_append).to eq ""
+        expect(property.de.description_append).to eq ""
+        expect(property.es.description_append).to eq ""
+        expect(property.zh.description_append).to eq ""
+        expect(property.zh_tw.description_append).to eq ""
+      end
+    end
+  end
+
   describe "#add_license_number" do
 
     let(:property) { Roomorama::Property.new('test') }
-    before { described_class.new.send(:add_license_number, property, response) }
+    before {
+      property.description_append ||= ""
+      property.de.description_append ||= ""
+      property.es.description_append ||= ""
+      property.zh.description_append ||= ""
+      property.zh_tw.description_append ||= ""
+
+      described_class.new.send(:add_license_number, property, response)
+    }
 
     context "when license is empty" do
       let(:response) {
@@ -232,11 +299,11 @@ RSpec.shared_examples "Waytostay property client" do
         )
       }
       it "should not add description and translations" do
-        expect(property.description_append).to be_nil
-        expect(property.de.description_append).to be_nil
-        expect(property.es.description_append).to be_nil
-        expect(property.zh.description_append).to be_nil
-        expect(property.zh_tw.description_append).to be_nil
+        expect(property.description_append).to eq ""
+        expect(property.de.description_append).to eq ""
+        expect(property.es.description_append).to eq ""
+        expect(property.zh.description_append).to eq ""
+        expect(property.zh_tw.description_append).to eq ""
       end
     end
 
@@ -249,11 +316,11 @@ RSpec.shared_examples "Waytostay property client" do
         )
       }
       it "should not add description and translations" do
-        expect(property.description_append).to eq "License number: ABC"
-        expect(property.de.description_append).to eq "Lizenznummer: ABC"
-        expect(property.es.description_append).to eq "Número de licencia: ABC"
-        expect(property.zh.description_append).to eq "许可证号: ABC"
-        expect(property.zh_tw.description_append).to eq "許可證號: ABC"
+        expect(property.description_append).to eq "License number: ABC\n"
+        expect(property.de.description_append).to eq "Lizenznummer: ABC\n"
+        expect(property.es.description_append).to eq "Número de licencia: ABC\n"
+        expect(property.zh.description_append).to eq "许可证号: ABC\n"
+        expect(property.zh_tw.description_append).to eq "許可證號: ABC\n"
       end
     end
   end